launchpad_loggerhead: Fix robots.txt
On Python 3, the response to a request for `robots.txt` needs to be
bytes or Gunicorn rejects it.

We didn't notice this on production or staging because this file is
served via an Apache `Alias` directive rather than going via the
appserver, but fixing it seems to make sense anyway.