Added new logic to filter BugNotificationRecipient
filters_by_person (dictionary) is the new data structure introduced:
key is the person_id and the value is a set of tuple with the following format (BugNotificationRecipe, filter_id).
Thanks to that structure we can filter the muted notifications simply by removing them from the filters_by_person[person_id] set.
The remaining notifications will be sent to the user.

LP: #2019428