Refactor time zone name handling in DateTimeWidget
`DateTimeWidget` currently works out the time zone name to display by
calling `.tzname(None)` on a time zone object.  This works with `pytz`,
but it won't work once we switch to `dateutil.tz`, and there doesn't
seem to be a reasonable way to get it to work with that time zone
provider without poking about in private attributes; the best we could
do would return a time zone abbreviation string rather than the original
time zone name the user asked for (e.g. "CET" rather than
"Europe/Paris"), which isn't very satisfactory.

Since the whole approach of expecting to be able to recover the original
time zone name from a time zone object seems fundamentally dodgy, a
better approach is to explicitly keep track of the original time zone
name to start with.  This needs some changes to `LaunchBag`, but the
result should be much more robust and can easily be made to work with
any time zone provider.