Document database role management for charms
When setting up charmed instances of the librarian recently, I had to
invent some approaches to handling database roles.  I expect much the
same approaches to be used for a number of other charms, so I wanted to
at least write down what I did so that other people don't need to
reinvent it if they find themselves doing the same thing.