Update distro from 16.4 to 18.4 in starter docs
We ask users to create a 16.4 container in the started docs, which more often than not leads to issues (e.g. the IPv4 of the container not being available). This is not an issue with ubuntu 18.4 as far as I've seen it, so it will make the setup process easier and more error proof.