charm: Expand documentation of librarian volume handling
The documentation of how to set up a Ceph volume on the librarian was a
bit brief.  Expand this based on what I just did on qastaging.

(At some point it may be worth investigating whether Juju's built-in
storage handling is robust enough to replace this; but this is workable
for now.)