Add region to builder metrics
Launchpad builders don't really have a built-in notion of "region" (i.e.
a collection of builders that are on common infrastructure and that are
all managed as a group).  However, in practice, we always create
builders with names that allow us to infer a region from their names:
for instance, `lcy02-amd64-{001..120}` currently form a region.

Add the region to builder metrics where appropriate, so that we can
write alerts of the form "no builder in a given region has been in the
idle or building state for the last hour", which would indicate a
problem with that region's builder reset handling.