charm/launchpad-librarian: Add http interface support
This allows relating `launchpad-librarian` to a deployment of the
`haproxy` charm to have it load-balance the various workers and expose
them on the ports expected by other parts of Launchpad.

The code is based on the `ols-http` layer, but I had to write custom
code for it because the librarian is rather special in a few ways: it
has four different sets of ports with different purposes (download,
upload, restricted download, and restricted upload), and multiple
workers per set; and the current production `haproxy` configuration uses
somewhat different options for the download vs. upload workers.

While this doesn't result in the exact same `haproxy` configuration that
we have on production, I think it's equivalent.