charm/launchpad-librarian: Restart when worker count is changed
If the `workers` config option is changed in isolation, then none of the
files currently checked for changes by the `configure` hook are changed,
so it fails to restart the service to start up the new workers.  Avoid
this problem by adding
`/lib/systemd/system-generators/launchpad-librarian-generator` to the
list of monitored files.