charm/launchpad-appserver: Add http interface support
This allows relating `launchpad-appserver` to a deployment of the
`haproxy` charm to have it load-balance the various workers and expose
them on the ports expected by other parts of Launchpad.

The code is based on the `ols-http` layer, but I had to write custom
code for it because the appserver exposes two ports with different
purposes (the main application, and XML-RPC); and the current production
`haproxy` configuration uses slightly different options for those two,
with special error page handling for the main application.

While this doesn't result in the exact same `haproxy` configuration that
we have on production, I think it's equivalent.