charm: Update ols-layers
This pulls in a number of improvements.  The most relevant ones are:

  f157081b16 Remove libpq-dev, python-dev, gcc from the ols-pg packages

In combination with some changes to `build-snaps`, `build-packages`, and
`reactive-charm-build-arguments` in our `charmcraft.yaml` files, this
allows us to build charms that contain wheels rather than sdists in
their `wheelhouse` directory for dependencies of charm code itself, and
to drop the requirement for the above packages to be installed in most
of our deployed machines; this makes all our charms be a bit faster to
deploy and use a little less disk space.

  39c38b1f12 Add tarball_payload_name layer option for multi-charm services

This was my motivation for this update; I wanted to make
`launchpad-assets` use the `ols-http` layer for `haproxy` integration,
but ran into the `service_name` layer option being used for both
deciding which artifact to download from Swift and naming the `haproxy`
frontend.  This change will let us decouple those two things.