charm/launchpad-assets: Add http interface support
Since this charm is fairly simple and only exposes a single port, we can
just let the `ols-http` layer do most of the work.  The only downside is
that this involves renaming the `port_assets` configuration key to
`port`, but that just means we'll need to change the configuration key
in the Mojo spec when updating the charm revision.

I added the most minimal possible `/_status/check` URL to support
`ols-http`'s `nrpe` relation.