Fix update-pkgcache crash due to binary deletion
The `update-pkgcache` script started crashing on production about a week
ago, and mysteriously started working again today without us changing
anything.  Based on code inspection, the only way I can see for this
particular crash to happen is if a binary was deleted between
`DistroSeriesPackageCache.findCurrentBinaryPackageNames` and
`DistroSeriesPackageCache._update` (`update-pkgcache` is slow enough
that it makes many commits along the way, so this is possible), and
simulating that in a test reproduces the same crash.

It seems safe to skip the affected binary in such cases, since
`DistroSeriesPackageCache.removeOld` will eventually clean up its cache
entries.

LP: #2023698