Always pretend that "source packages" for Conda have no files
`CIBuildUploadJob.format_by_repository_format` is supposed to ensure
that no files are attached to Conda source packages, which are intended
only as skeletons to fit into the rest of Launchpad's data model so that
things like the web UI and copy operations work.  However, there appears
to be exactly one case on production where a file somehow got attached
to a source package in a Conda-format archive, and the attempt to reap
that file from the archive is failing because we can't compute its path.
That file isn't actually published in Artifactory, so skipping it
shouldn't be a problem, but we do need the reaper not to crash.

Since Conda "source packages" should never have any files, just enforce
this in `SourcePackagePublishingHistory.files`.  That's slightly less
awkward than the existing workaround in `ArchivePublisherBase.publish`,
and it ensures that we handle this case both on publication and on
removal.

I've left an XXX comment in place because I think we should remove this
once we figure out how that file got attached in the first place and
ensure that that doesn't happen again.  We'd also need to ensure that
this query returns no rows:

    SELECT spph.id, lfa.filename
    FROM
        sourcepackagepublishinghistory spph,
        sourcepackagereleasefile sprf,
        libraryfilealias lfa
    WHERE
        spph.archive IN (
            SELECT id FROM archive WHERE repository_format = 2
        )
        AND spph.sourcepackagerelease = sprf.sourcepackagerelease
        AND sprf.libraryfile = lfa.id
        AND spph.dateremoved IS NULL;

LP: #2023315