Fix database initialization in charmed deployments
In charmed deployments, `launchpad-admin` is typically related to
`postgresql`, which causes the `postgresql` charm to automatically
create an empty `session_dev` database.  This confuses the `create`
target in `database/schema/Makefile` into assuming that the
`session_dev` database has already been initialized and not doing so
again.

I couldn't find any record in git history of why we bother to perform
this check for `session_dev` when we don't do so for any of the other
databases.  I think it makes more sense to assume that if you're running
the `create` target then you mean to actually create all the databases,
so this commit brings the handling of `session_dev` into line with all
the other databases.