charm: Fix names of convoy symlinks for staging
On staging, the build label includes a "-db" suffix (in order that we
can have a separate stream of deployment artifacts built from the
`db-devel` branch); that suffix is present in some directory names, but
the `+combo` URLs generated by the appserver just use the bare commit ID
and not the suffix.  Adjust the convoy symlinks to account for this.