charm: Don't set Cache-Control header on +apidoc
This location is used to serve machine-readable API documentation used
by `lazr.restfulclient`, and if it sees a response with `Cache-Control:
public,max-age=5184000` then it won't refresh its cache until `max-age`
has passed - 60 days!  This is OK for URLs whose contents change very
rarely, but for machine-readable API documentation we want changes to be
effective more or less immediately, and they aren't otherwise hit often
enough to need such aggressive caching.

(On production, https://launchpad.net/+apidoc/ and
https://api.launchpad.net/ previously had different caching
configuration for the same documentation.  This brings the former into
line with the latter, since they're now routed to the same backend, or
at least will be once we revert an emergency frontend patch - but I
think that's OK.)