charm/launchpad-scripts: Add start-services/stop-services actions
In our legacy production deployment, we stop and start these services
around certain kinds of maintenance actions, mainly database schema
upgrades.  The preferred way to do this sort of thing in Juju is using
actions, so define a couple of suitable actions; once we deploy this to
production, we'll set up an SSH trigger allowing developers to run any
of an allowlist of actions on production.