Remove lp.services.scripts.tests.run_script
It turned out that the only way in which
`lp.testing.script.run_script`'s `env` argument was ever used was to
supplement the environment of the calling process, so I changed it to
use the simpler `extra_env` calling convention from
`lp.services.scripts.tests.run_script`.  Aside from that, the main
difference was that `lp.services.scripts.tests.run_script` tested the
exit code itself, and it's probably clearer to have that be explicit in
tests anyway.

(Note that some tests already had their own separate assertion of the
exit code, although it isn't always visible in the patch context.)