Fix bugs in conversion of LoginToken to Storm
`LoginTokenSet.new` needs to explicitly add the new token to the store,
as it may not be linked to any other object if it doesn't have a
requester (e.g. `LoginTokenType.BUGTRACKER` tokens).

`test_LoginTokenPruner` needs to flush SQL statements for new tokens to
the database before trying to get their IDs.