Fix bugs-emailinterface.rst after conversion of Message to Storm
The `DecoratedResultSet` used in `Bug.messages` to replace the previous
`SQLRelatedJoin(prejoins=["owner"])` can't support `__contains__`
without a fair bit of work.  Fortunately this wasn't needed very much.
There was one use in `Bug.linkMessage` which I already replaced with an
explicit Storm query, but I missed the uses of it in
`bugs-emailinterface.rst`.

In a test we don't need to care much about performance, so I just used
`set(bug.messages)` to materialize the result of the query and turn it
into something that supports `__contains__`.