Export ISSHKey.getFullKeyText
There are currently several cron jobs on loganberry that export email
and SSH key data for the members of several teams, using
`scripts/list-team-members`; this information is used as part of various
integrations such as lists.ubuntu.com posting privileges, the ability
for Ubuntu members to upload files to people.ubuntu.com using SFTP, and
so on.  However, doing this with ad-hoc cron jobs on a Launchpad machine
isn't ideal; it would be better if it were possible to get this
information over the API, and then IS would only need a
suitably-privileged bot account.

As far as I can see, the only API gap here is that we can only extract
the raw key text for an SSH key, and not the "full" key text that's
decorated with the key type and comment in a format suitable for adding
to an OpenSSH `authorized_keys` file.  That transformation isn't quite
trivial, so export a method that does it on the webservice.