Convert lp.buildmaster.queuedepth queries to Storm expressions
We generally prefer having less in the way of constructing SQL queries
by pasting strings together.  This will also make it a little easier to
convert `Processor` to Storm.

Most of the transformations are reasonably obvious; I switched from
`EXTRACT(EPOCH FROM ...)` to `date_part('epoch', ...)` because the
latter can be expressed in Storm expressions more conveniently, but
https://www.postgresql.org/docs/10/functions-datetime.html says that
they're equivalent.