Start feature controller earlier in Twisted services
After migrating buildd-manager to a charmed deployment, it quite
frequently seems to start up in a bad state, where all its database
operations fail with errors like "permission denied for relation
builder".

I can reproduce this on staging, but only sometimes, and unreliably: it
seems to be a race condition of some kind.  My best guess is that it has
something to do with the different database connection arrangements for
charmed deployments, where we switch to the desired database role after
connecting.  As such, I guessed that setting up the feature controller
earlier in the service startup sequence might help, since that makes a
database connection, and it would make some sense that it might matter
whether this happens before or after we start a bunch of threads; and
indeed when I try that I don't seem to be able to reproduce the problem
on staging any more.

This isn't entirely satisfactory, and I'd like to understand the problem
better, but perhaps this will do for now.