charm: Set up frontend relations for the librarian
The Apache configuration files here are lifted fairly directly from
production, although qastaging and staging are much the same.  The
`service_name` settings are passed through the stack of Juju relations
(with Squid and Apache) and can be used in Apache `balancer://` URLs
with no extra configuration required.

At the moment, the public and restricted librarians are on different
sets of IP addresses.  I _think_ that this was mostly due to historical
issues around issuing TLS certificates that are no longer an issue, so
I'm going to try to get things working with a single set of public IP
addresses on (qa)staging; if that doesn't work out then I'll fall back
to splitting the `vhost-config` relation into two and deploying two
Apache applications, but it doesn't end up being all that fundamentally
different.