Introduce a new GATHERING build status
If a build has large output files, it can take buildd-manager a long
time to download those from the builder.  This happens in a single long
scan cycle for the builder in question, and while it's in progress
there's almost no visual indication of that fact; the only way to tell
is that the build is in the BUILDING state with a build log (not just a
log tail) attached to it.

More importantly, if somebody decides that a build is hung when it's in
that state and tries to cancel it, the cancel attempt is largely ignored
(there's currently no way to interrupt the downloads in progress) but
when the download completes buildd-manager will raise `AssertionError:
Can't change build status from CANCELLING to UPLOADING.`

To avoid this problem, introduce a new status called `GATHERING`, which
builds pass through on their way from `BUILDING` to `UPLOADING`.  Builds
in this state won't show up as cancellable, and it also makes the true
state a little more visible to users.