charm: Set up frontend relations for the appserver
The Apache configuration files here are derived from those on
production, with added templating and some minor tidying-up.  The
`service_name` settings are passed through the stack of Juju relations
(with Squid and Apache) and can be used in Apache `balancer://` URLs
with no extra configuration required.

The balancer arrangements do present a slight problem with appserver
requests that explicitly need to be uncached, since the
`squid-reverseproxy` charm passes through the services it receives from
its `website` relation with the same names.  We work around this by
adding a `cached-launchpad-appserver-main` service that explicitly goes
through the cache, and once
https://code.launchpad.net/~cjwatson/squid-reverseproxy-charm/+git/squid-reverseproxy-charm/+merge/447787
and
https://code.launchpad.net/~cjwatson/squid-reverseproxy-charm/+git/squid-reverseproxy-charm/+merge/448080
land we'll configure the `squid-reverseproxy` charm to only re-export
selected services from its `website` relation.  A downside of this is
that we need an extra haproxy frontend with its own port, but this is
only really visible within the frontend stack and I don't think it
should pose an operational problem.

The "api" virtual host is published on a separate relation because it
currently needs to be on a separate IP address in order to support old
Ubuntu versions whose Python versions didn't support SNI.  On production
we currently have manually-configured additional IP addresses on the
frontends to achieve this; in a charmed deployment, the simplest
approach is just to deploy a separate Apache frontend application for
each set of public IP addresses we want to have.

There are a number of other virtual hosts on production that don't proxy
to the appserver, because they only serve redirections or static files;
there are also a number of references in this merge proposal to
`DocumentRoot`s that don't exist yet.  My plan is to handle those with a
separate charm that can be a subordinate to the relevant frontend and
can ship those simple virtual hosts and additional files.