Silence some more librarian-gc OOPSes on non-production systems
The errors about data being missing from disk or from Swift mostly don't
make sense if we have an upstream host (and in one case the attempt to
suppress this error no longer worked in a charmed deployment).  There
are over 160000 such OOPSes daily from qastaging right now, so it's
worth reducing the load on the error-processing system a bit.