charm: Separate restricted librarian vhost configuration
I'd hoped to avoid this and to be able to combine the public and
restricted librarians onto a single set of IP addresses, but the
restricted librarian relies on wildcard DNS
(*.restricted.launchpadlibrarian.net etc.), and Let's Encrypt only
supports wildcard SANs if you're using the DNS-01 challenge type, which
I don't think we can easily arrange in our environment.  As such, we'll
need separate frontends for each of the public and restricted
librarians, which means having a separate implementation of the
`vhost-config` interface for each of them.