charm: Fix final part name
The example in
https://juju.is/docs/sdk/charmcraft-yaml#heading--the-reactive-plugin
calls the final (and in that case only) part "charm".  I'd thought that
was just an example.  However, as of
https://github.com/canonical/charmcraft/commit/35adadf42c625c5ee5ed6d611c9bdec10413ae59
in charmcraft 2.4.0, calling the final part anything else causes charms
to be misbuilt; they end up containing unmodified copies of the
top-level `actions.yaml`, `config.yaml`, and `metadata.yaml` files
rather than merging the information provided by each layer.