Commit after running SET ROLE
In charmed deployments configured with `set_role_after_connecting:
True`, we execute a `SET ROLE` command immediately after connecting to
the database.  However, according to
https://www.postgresql.org/docs/10/sql-set.html which is referenced by
https://www.postgresql.org/docs/10/sql-set-role.html, the effects of the
`SET` command disappear if the transaction is rolled back.  We must
therefore commit the transaction to ensure that the effects persist
until the end of the session.

I strongly suspect that this is why we've been seeing `buildd-manager`
sometimes running as the wrong database role: it's to be expected that
Launchpad code will roll back transactions sometimes, for example as
part of handling an exception, and if that happened to be one of the
first things that `buildd-manager` did then we'd see this effect.