charm: Add offline mode support
Our frontends show a custom error page if backends are unavailable.
Some of our legacy deployments support switching between different error
pages in these cases, which is useful for telling users about scheduled
maintenance.  Add an `offline_mode` configuration option for this; I
expect to make use of this in the `staging-restore` script.