Add RevisionStatusReport.distro_arch_series
An awkward legacy of the order in which revision statuses and CI builds
were designed is that revision status reports linked to a CI build don't
record the series and architecture used for each individual CI build
job.  This is a problem when we need to be able to do things like filter
the artifacts produced by a CI build to select only those built on a
particular series, and in general it seems like information that would
be worth tracking in the database.  Add a column for it.