Track the series/architecture for each revision status report
An awkward legacy of the order in which revision statuses and CI builds
were designed is that revision status reports linked to a CI build don't
record the series and architecture used for each individual CI build
job.  This is a problem when we need to be able to do things like filter
the artifacts produced by a CI build to select only those built on a
particular series, and in general it seems like information that would
be worth tracking in the database.

Model the new `RevisionStatusReport.distro_arch_series` column for this,
allow passing it when creating a new report, and set it when requesting
CI builds if we can.  That last step has to hardcode Ubuntu for the same
reasons that `lp.code.model.cibuild.determine_DASes_to_build has to do
so; and as usual we have to keep a clear head about the distinction
between the series/architecture that runs the build farm job (i.e. the
container where `lpci` itself is run) and the series/architecture of
each individual CI build job (i.e. the container started by `lpci` that
runs the workload).