Restrict CIBuildUploadJob to binaries built for the target series
The `CIBuildUploadJob` mechanism (used by `Archive.uploadCIBuild`)
presumes that it makes sense to upload all the binaries produced by a
given CI build to the same context.  However, it seems that for the
project where we're using this most heavily at the moment, the security
team would strongly prefer to build both focal and jammy binaries from
the same CI build (using `series: focal` and `series: jammy` in
individual CI jobs), reducing the number of git branches they need to
maintain.  For this to work properly, we need to filter the uploads:
it's unlikely to make sense to publish Python 3.8-based wheels to jammy,
or Python 3.10-based wheels to focal.

Making this work requires tracking the series and architecture for each
"revision status report" (corresponding to individual CI jobs, or to
rows displayed in the web UI under each revision).  With that in place,
we can skip binaries built for the wrong series.

There's an additional wrinkle: we currently rely on being able to build
on (e.g.) Ubuntu focal and upload the results to (e.g.) <private
distribution> focal.  These series unfortunately don't have anything
formally linking them other than the name, so right now the only option
seems to be to compare series names.  This is an unpleasant hack, but
fortunately `CIBuildUploadJob` isn't particularly deep in the data model
so it should be viable to fix this later.