Flush store before extracting job state
Since converting `Job` to Storm in c69d6205ab, the store is no longer
automatically flushed when the `id` column is fetched, so we have to be
a little more careful to ensure that it's flushed at appropriate times.
Some Celery tasks are currently showing up with their job IDs as None,
indicating that this isn't always being done properly.  This appears to
be because the before-commit hook that extracts the job state runs
before the Storm hook that flushes the store at the start of a commit,
so if the statement that created the `Job` row hadn't yet been flushed
for some other reason then we might not know the job ID yet.

The test suite evaded this problem because Celery-based job tests
typically use `FeatureFixture`, which causes
`BaseRunnableJob.celeryRunOnCommit` to make another database query after
creating the job, thus causing an implicit flush.

Flushing the store before extracting the job state should be a complete
fix for this class of problem.