charm/launchpad-admin: Run pgkillactive/pgkillidle
On production, we run `pgkillactive` and `pgkillidle` to control
connections that hang around for too long.  (The `postgresql` charm has
a similar facility, but it isn't fine-grained enough for us.)  This
currently runs on bare metal and needs to be moved into a charm.  The
`launchpad-admin` charm seems to be a reasonable choice: it has database
admin credentials and a copy of the Launchpad tree, and we can easily
set up a cron job exactly the way we want.

We have to do some regex-mangling to match `pgkillactive`'s current
interface; I thought it better not to block moving this function into a
charm on improving that interface.