Silence an error from "make -C database/replication stagingsetup"
The weekly staging restore process currently includes the following
output:

  Setting feature flags
  psql -d lpmain_staging -c "INSERT INTO featureflag (flag, scope, priority, value) VALUES ('profiling.enabled', 'team:launchpad', 0, 'on')"
  INSERT 0 1
  psql -d lpmain_staging -c "INSERT INTO featureflag (flag, scope, priority, value) VALUES ('librarian.swift.enabled', 'default', 0, 'on')"
  ERROR:  duplicate key value violates unique constraint "feature_flag_pkey"
  DETAIL:  Key (scope, flag)=(default, librarian.swift.enabled) already exists.
  Makefile:57: recipe for target 'stagingsetup' failed
  make: [stagingsetup] Error 1 (ignored)

This is in fact harmless (the leading `-` on those lines means that we
carry on anyway), but it confused me briefly.  I think it's a little
neater to use `ON CONFLICT DO NOTHING` to ensure that the statement
succeeds even if an identical feature rule already exists in the data
synced from production.