Expand relative directories in PublisherConfig.root_dir
The `PublisherConfig` table is designed quite strangely: it includes a
`root_dir` column which stores the root file system path at which each
distribution is published.  This is a problem for charmed deployments of
the primary Ubuntu and copy archive publishers, because our charms don't
use the same base directory as our legacy deployments.

Fortunately, all the existing paths are absolute, so we have a
reasonable migration path that doesn't require us to change the database
at the same time as migrating to a new deployment.  Interpret relative
paths in `PublisherConfig.root_dir` as being relative to a new
`config.archivepublisher.archives_dir` configuration entry; we can set
this to appropriate values in both legacy and charmed deployments, and
then change existing production database rows to use relative paths.