charm: Set up Apache relation for ftpmaster-publisher
Apache needs to run on the same machine as the publisher so that it can
serve the published data.  The easiest way to do that is to convert the
publisher into a subordinate charm and implement the `apache-website`
relation, which will allow us to deploy it in conjunction with the
standard `apache2` charm.

The virtual host configuration here is based on what's currently in
production.