Pass SITE_NAME environment variable to run-parts
A number of the hook scripts in
https://bazaar.launchpad.net/+branch/ubuntu-archive-publishing rely on
being able to tell the difference between our different deployments, so
that they can behave differently in production and non-production
environments.  For example,
https://bazaar.launchpad.net/+branch/ubuntu-archive-publishing/view/head:/publish-distro.d/10-sign-releases
uses different signing keys for different environments, since of course
dogfood doesn't have the production keys.

At the moment, we rely on the `LPCONFIG` variable (implicitly inherited
by these hook scripts from the parent process) for this.  However, in a
charmed deployment, `LPCONFIG` is set purely based on the charm name,
and is the same across production and non-production deployments of the
same charm.  As a result, we can't rely on it for this any more.

An approach that we've taken in a few other places is to conditionalize
behaviour based on the main site's hostname.  This has the advantage
that the condition is spelled using an existing user-visible name (e.g.
"launchpad.net") rather than an internal codename that you're just
expected to know (e.g. "production" or the various values of
`LPCONFIG`).  To support this, I've invented the `SITE_NAME` environment
variable, which the publisher now sets to the main site's hostname when
running its hook scripts.