Refactor queries in PublishingSet.getBuildsForSourceIds
`PublishingSet.getBuildsForSourceIds` made one database query to work
out the set of build IDs (in an inefficient way that pulled a lot more
data back into Python than needed), and then a second query to get the
full result set to return.  Refactor this to do all the work in a single
query.

The approach I've taken here also moves the `BinaryPackageBuild` parts
of the query into a left join, which should soon make it easier to
extend this method to support CI builds as well.