Make StormBase implement IDBObject
According to https://bugs.launchpad.net/launchpad/+bug/591841, there was
once a problem with getting non-`SQLBase` Storm models to implement
`IDBObject`, and as a result the `IPrimaryObject` adapter had to be
registered on the very broad `Interface` rather than on the more
sensible `IDBObject`.

Nowadays, all our Storm models use `StormBase` (enforced by lint rules),
and there doesn't appear to be any problem with declaring that that
class implements `IDBObject`, so we can replace this ancient hack with
something neater.