charm: Drop restricted-vhost-config relation
I'd originally separated these because it looked like we might need to
use separate Apache units for the public and restricted frontends for
TLS certificate handling reasons, but we later decided against that.  It
turns out that trying to have two `apache-vhost-config` relations
connected to the same Apache application doesn't work: they both get
written to file names derived from the `servername` configuration option
on the Apache application, and so one overwrites the other.

Since we don't actually need the flexibility of having the public and
restricted librarian be published on separate Apache units, we can solve
this problem by collapsing them into one relation.