Convert Product to Storm
There are a number of places that try to handle either a `Distribution`
or a `Product` using similar code, so to avoid too much hassle I also
converted `Distribution.owner` and `Distribution.driver` to the new
style.  For similar reasons I ended up changing
`SQLObjectVocabularyBase.__contains__` to accept `Storm` objects as well
as `SQLBase` objects and to use Storm-style query syntax; in some ways
this is a little odd, but `SQLObjectVocabularyBase` only has a few users
left and will go away soon.