charm: Add librarian-logs rsync module to librarian frontend
The `launchpad-scripts` charm needs to be able to rsync librarian logs
from the appropriate frontends so that it can parse them and update
download counts.  On the legacy frontends this is configured manually;
on charmed frontends we need to configure it in code somewhere.

To avoid having to fork the `apache2` charm, we add a custom
`librarian-logs` subordinate to `launchpad-frontend-extras`.  This can
be related to the librarian frontend application to configure an rsync
module there.