charm: Add launchpad-cron-control
One of the remaining odd pieces of our legacy frontend stack is
"cron-control": a mechanism to centrally enable and disable cron jobs,
checked by `lp.services.scripts.base.cronscript_enabled`.  This is an
important part of some of our rollout procedures, such as database
schema migrations.  It's worth having this mechanism in addition to
things like `stop-services` and `start-services` actions on individual
charms, because it gives us a very quick and easy central switch.

The actual machinery behind this amounts to a simple script to
manipulate a `.ini` file which is served internally over HTTP, so it's
pretty easy to get an equivalent up and running using charms.  While the
original script can be found at
https://bazaar.launchpad.net/+branch/canonical-mojo-specs/view/head:/launchpad-manual-servers/files/croncontrol-ctl.sh,
I thought it would be more readable to reimplement similar logic in
Python rather than incorporating that shell script directly.  Relying on
Juju actions as the primary interface to this also means there's no
particular need to worry about locking.

Using this in practice for a database deployment might look something
like this:

  $ juju run-action --wait launchpad-cron-control/leader disable-cron job=publish-ftpmaster
  [...]
  $ curl http://cron-control.launchpad.test/cron.ini
  [DEFAULT]
  enabled = True

  [publish-ftpmaster]
  enabled = False

  # wait for slow Ubuntu archive publisher job to complete

  $ juju run-action --wait launchpad-cron-control/leader disable-cron-all
  [...]
  $ curl http://cron-control.launchpad.test/cron.ini
  [DEFAULT]
  enabled = False

  # wait for any remaining slow jobs to quiesce

  # run migration

  $ juju run-action --wait launchpad-cron-control/leader enable-cron-all
  [...]
  $ curl http://cron-control.launchpad.test/cron.ini
  [DEFAULT]
  enabled = True