Bump buildd-manager open file limit some more
We're bumping into the current limit on production, and I think it's
because a few things have changed since we last set that limit: we have
more active builders, and we make a few more outgoing connections than
we did then (particularly to issue tokens for private builds).  Make the
limit more generous.