daemons/buildd-manager.tac: Drop RLIMIT_NOFILE fiddling
When I originally added code to change `RLIMIT_NOFILE` here, the intent
was to _raise_ the limit from the relatively low OS default.  However,
since then, we've switched to running `buildd-manager` from a systemd
service deployed via a charm, and
`charm/launchpad-buildd-manager/templates/launchpad-buildd-manager.service.j2`
sets `LimitNOFILE=65536`; as a result, this code now has the effect of
_lowering_ the limit.

Let's just rely entirely on the configuration in the systemd service.