charm: Add launchpad-db-update
This handles database schema updates, adding `preflight` and `db-update`
actions.  It should allow us to replace another function for which we
currently rely on non-Juju infrastructure.

On production, we can't do this from `launchpad-admin`, since the
distinction between the `master`/`stable` and `db-devel`/`db-stable`
branches means that we need to have different commits deployed for
ordinary administrative functions and for schema updates.  I considered
making the behaviour of this charm be optional behaviour in
`launchpad-admin` rather than adding a new charm, but I ended up needing
to turn most of `launchpad-admin` off, and if the two modes are going to
have mostly disjoint charm code anyway then they might as well have
different charms.