charm: Fix queue permissions in ftpmaster-uploader
This syncs this charm up a bit more closely with
`launchpad-ppa-uploader`.  Previously, the queue directories were all
owned by `root:root` so neither `txpkgupload` nor `process-queue` could
do anything in them; in addition, the `launchpad` user needs to be a
member of the `txpkgupload` group so that it can see uploaded files.