charm: Add launchpad-postgresql-extras
The current production databases have a few unusual quirks, applied
using the `ols-postgresql-extras` subordinate charm.  This hasn't been
open-sourced and has quite a lot of cruft, so I thought it would be best
to have our own modernized version of it.  This charm should be suitable
for use in all environments.

When related to the `postgresql` charm, this adds the following
features:

 * Installs `launchpad-database-dependencies` (thereby pulling in
   various extensions we need).
 * Installs and configures `pgbouncer`, optionally reporting metrics to
   `telegraf` (requires
   https://code.launchpad.net/~cjwatson/charm-telegraf/+git/charm-telegraf/+merge/450883).
 * Optionally installs a cron job to push full database backups to
   another machine, which can be used for the weekly sync of staging
   from production.