Improve handling of Firefox subprocesses in tests
Firefox is run via Selenium by some test cases.  Since it's rather
heavyweight, we run it in a layer so that it can be started up and shut
down just once for a group of test cases.

I happened to notice recently that Firefox wasn't actually being shut
down correctly, and continued to run long after any test cases that used
it had finished.  This turned out to be because the `YUITestLayer`
teardown process was calling `selenium.webdriver.Firefox.close` (which
closes the current window) rather than `selenium.webdriver.Firefox.quit`
(which quits the driver and closes every associated window).

In the process, I also noticed that `lp.testing.tests.test_html5browser`
started a separate browser for each individual test case, which seemed
inefficient.  I've split out a new `WebBrowserLayer` to fix this, which
can be used to run tests of just the web browser part without bringing
in the extra baggage of `FunctionalLayer` the way that `YUITestLayer`
does.

I've verified that the process management now works correctly by running
`while :; do date; ps xf; sleep 2; echo; done` in parallel with the test
suite, and checking both that it uses a common browser instance and that
it cleans it up when the relevant layer is torn down.